chore(sdk/java): generate code
Using the new sdk/java/dagger-java-sdk/src/gen output directory
instead of the one under a target directory.
That way all the generated types are serialized. This might help to
create SDK jars but also to see the impact of any change as the generated
types will be available.

Fix runtime so generated code is module-related only

Signed-off-by: Yves Brissaud <gh@lgtd.io>