sdk/go: implement moduleTypeDefs
Go SDK now has a moduleTypeDefs method that is dedicated to return
types exposed by the module.

The `moduleTypeDefs` is similar to `moduleRuntime` but the generated
entrypoint only contains the types and not the invocation code.

Codegen contains a new flag `--typedefs-only` to only generate the
part related to types.

Other than this flag, one change in the codegen is a `register` function
is created to handle the management of the types. `invoke` function
will call `register` when the parent name is empty.
And basically when `--typedefs-only` is set, the invoke function is not
created but the `register` is and is called instead of `invoke`.

Signed-off-by: Yves Brissaud <gh@lgtd.io>